without dtos

+++ src/main/java/com/sgkhmjaes/jdias/service/impl/ConversationDTOServiceImpl.java
@@ -2,14 +2,15 @@
 package com.sgkhmjaes.jdias.service.impl;
 
 import com.sgkhmjaes.jdias.domain.Conversation;
+import com.sgkhmjaes.jdias.domain.Message;
 import com.sgkhmjaes.jdias.domain.Person;
-import com.sgkhmjaes.jdias.domain.User;
 import com.sgkhmjaes.jdias.repository.ConversationRepository;
+import com.sgkhmjaes.jdias.repository.MessageRepository;
 import com.sgkhmjaes.jdias.repository.PersonRepository;
 import com.sgkhmjaes.jdias.repository.UserRepository;
 import com.sgkhmjaes.jdias.repository.search.ConversationSearchRepository;
+import com.sgkhmjaes.jdias.repository.search.MessageSearchRepository;
 import com.sgkhmjaes.jdias.security.SecurityUtils;
-import com.sgkhmjaes.jdias.service.dto.AuthorDTO;
 import com.sgkhmjaes.jdias.service.dto.ConversationDTO;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -20,9 +21,12 @@ import java.util.stream.Collectors;
 import java.util.stream.StreamSupport;
 import static org.elasticsearch.index.query.QueryBuilders.*;
 import java.time.LocalDate;
-import java.util.HashSet;
 import java.util.Set;
 import java.time.ZonedDateTime;
+import java.util.Collections;
+import java.util.Comparator;
+import java.util.UUID;
+import org.hibernate.Hibernate;
 
 @Service
 @Transactional
@@ -32,60 +36,86 @@ public class ConversationDTOServiceImpl {
     private final ConversationRepository conversationRepository;
     private final ConversationSearchRepository conversationSearchRepository;
     private final UserRepository userRepository;
-    private final AuthorDTOServiceImpl authorDTOServiceImpl;
     private final PersonRepository personRepository;
+    private final MessageRepository messageRepository;
+    private final MessageSearchRepository messageSearchRepository;
     //private final ConversationDTO conversationDTO;
 
-    public ConversationDTOServiceImpl(ConversationRepository conversationRepository, UserRepository userRepository, AuthorDTOServiceImpl authorDTOServiceImpl,
-            PersonRepository personRepository, ConversationSearchRepository conversationSearchRepository/*, ConversationDTO conversationDTO*/) {
+    public ConversationDTOServiceImpl(ConversationRepository conversationRepository, UserRepository userRepository, 
+            PersonRepository personRepository, ConversationSearchRepository conversationSearchRepository, 
+            MessageRepository messageRepository, MessageSearchRepository messageSearchRepository
+            /*, ConversationDTO conversationDTO*/) {
         this.conversationRepository = conversationRepository;
         this.conversationSearchRepository = conversationSearchRepository;
         this.userRepository=userRepository;
-        this.authorDTOServiceImpl = authorDTOServiceImpl;
         this.personRepository = personRepository;
+        this.messageRepository = messageRepository;
+        this.messageSearchRepository = messageSearchRepository;
         //this.conversationDTO = conversationDTO;
     }
     
     public Conversation save(Conversation conversation) {
+        //avatar
         log.debug("Request to save Conversation : {}", conversation);
-        conversation.setCreatedAt(LocalDate.now());
+        Person currentPerson = getCurrentPerson ();
         conversation.setUpdatedAt(ZonedDateTime.now()); 
-        User currentUser = userRepository.findOneByLogin(SecurityUtils.getCurrentUserLogin()).get();
-        Person currentPerson = personRepository.getOne(currentUser.getId());
-        //AuthorDTO authorDTO = authorDTOServiceImpl.findOne(currentUser.getId());
-        //conversation.setAuthor(authorDTO.toString());        
+        conversation.addParticipants(currentPerson);
+        if (conversation.getCreatedAt() == null) conversation.setCreatedAt(LocalDate.now());
+        if (conversation.getGuid()== null) conversation.setGuid(UUID.randomUUID().toString());
+        if (conversation.getAuthor()== null) conversation.setAuthor(currentPerson.getDiasporaId());
         Conversation result = conversationRepository.save(conversation);
         conversationSearchRepository.save(result);
         return result;
     }
 
     public List<Conversation> findAll() {
+
         log.debug("Request to get all Conversations");
-        //List<Conversation> allConversation = conversationRepository.findAll();
-        /*for (Conversation conversation : allConversation) {
-            conversation.
-        }*/
-        return conversationRepository.findAll();
+        List<Conversation> conversations = getCurrentPerson().getConversations();
+        Collections.sort(conversations, (Conversation c1, Conversation c2) -> c1.getUpdatedAt().compareTo(c2.getUpdatedAt()));
+        for (Conversation conversation : conversations) {
+            Hibernate.initialize(conversation.getParticipants());
+        }
+        return conversations;
     }
 
     public Conversation findOne(Long id) {
         log.debug("Request to get Conversation : {}", id);
-        //Conversation conversation = conversationRepository.findOne(id);
-        //conversation.get
-        return conversationRepository.findOne(id);
+        Conversation conversation = conversationRepository.findOne(id);
+        if (conversation.getParticipants().contains(getCurrentPerson ()))return conversationRepository.findOne(id);
+        else return new Conversation();
     }
 
     public void delete(Long id) {
         log.debug("Request to delete Conversation : {}", id);
-        conversationRepository.delete(id);
-        conversationSearchRepository.delete(id);
+        Conversation conversation = conversationRepository.findOne(id);
+        Set<Person> participants = conversation.getParticipants();
+        Person currentPerson = getCurrentPerson ();
+        
+        if (participants.remove(currentPerson)){
+            currentPerson.getConversations().remove(conversation);
+            personRepository.save(currentPerson);
+        }
+        
+        if (participants.isEmpty()){
+            for (Message message : conversation.getMessages()) {
+                messageRepository.delete(message.getId());
+                messageSearchRepository.delete(message.getId());
+            }
+            conversationRepository.delete(id);
+            conversationSearchRepository.delete(id);
+        }
     }
-
+    
     public List<Conversation> search(String query) {
         log.debug("Request to search Conversations for query {}", query);
         return StreamSupport
             .stream(conversationSearchRepository.search(queryStringQuery(query)).spliterator(), false)
             .collect(Collectors.toList());
+    }
+    
+    private Person getCurrentPerson (){
+        return personRepository.getOne(userRepository.findOneByLogin(SecurityUtils.getCurrentUserLogin()).get().getId());
     }
     
 }
